Pushing again incase it crashes
I have tried to make some changes to the savepoints and rooms colliders to make them recognize when a room gets instantiated on top of the savepoint but to no avail. Now I'm playing around again with the code to see if I can make it keep going until max rooms has been hit.